requirements.txt
fastapi
uvicorn[standard]
pydantic
langchain
langchain-openai
python-dotenv
pytest
pytest-asyncio
httpx